# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of Oris Blog API.",
    "title": "Oris Blog.",
    "version": "1.0.0"
  },
  "host": "demo-orisblog-backend.herokuapp.com",
  "basePath": "/api/v1/blog/",
  "paths": {
...
    "/comments": {
      "get": {
        "tags": [
          "comment-CRUD"
        ],
        "summary": "Fetch Comments.",
        "operationId": "GetCommentRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "AuthorID",
            "description": "Fetch Comments by authorid or commentid or both",
            "name": "authorid",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "CommentID",
            "name": "commentid",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getCommentsResponse"
          }
        }
      },
      "post": {
        "tags": [
          "comment-CRUD"
        ],
        "summary": "Create a comment.",
        "operationId": "CreateCommentRequest",
        "parameters": [
          {
            "description": "Required body parameter to create a comment",
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateCommentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/CreatePostResponse"
          }
        }
      }
    },
    "/comments/{commentid}": {
      "put": {
        "description": "Update Comment",
        "tags": [
          "comment-CRUD"
        ],
        "operationId": "updateCommentRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "CommentID",
            "description": "Comment to be updated",
            "name": "commentid",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated data of comment",
            "name": "Comment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/updateCommentResponse"
          }
        }
      },
      "delete": {
        "description": "Delete Comment",
        "tags": [
          "comment-CRUD"
        ],
        "operationId": "deleteCommentRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "CommentID",
            "description": "Comment to be deleted",
            "name": "commentid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/deleteCommentResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Comment": {
      "type": "object",
      "properties": {
        "Author": {
          "type": "string"
        },
        "DateCreated": {
          "type": "string",
          "format": "date-time"
        },
        "LikeCount": {
          "type": "integer",
          "format": "int64"
        },
        "Likes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "PostID": {
          "type": "string"
        },
        "Replies": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "RepliesCount": {
          "type": "integer",
          "format": "int64"
        },
        "Text": {
          "type": "string"
        }
      },
      "x-go-package": "blog/docs/comment"
    },
    "CreateCommentRequest": {
      "type": "object",
      "properties": {
        "authorid": {
          "type": "string",
          "x-go-name": "AuthorID"
        },
        "postid": {
          "type": "string",
          "x-go-name": "PostID"
        },
        "text": {
          "type": "string",
          "x-go-name": "Text"
        }
      },
      "x-go-package": "blog/docs/comment"
    },
...
    "CreatePostResponse": {
      "type": "object",
      "title": "CreatePostResponse represents body of CreatePost response.",
      "properties": {
        "postid": {
          "type": "string",
          "x-go-name": "PostID"
        }
      },
      "x-go-package": "blog/docs/post"
    },
...
    "StatusResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "blog/docs/follow"
    }
  },
  "responses": {
...
    "CreatePostResponse": {
      "description": "This text will appear as description of your response body.",
      "schema": {
        "$ref": "#/definitions/CreatePostResponse"
      }
    },
...
    "deleteCommentResponse": {
      "description": "Delete Comment",
      "schema": {
        "$ref": "#/definitions/StatusResponse"
      }
    },
...
    "getCommentsResponse": {
      "description": "Fetch comments response according to supplied query params.",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Comment"
        }
      }
    },
...
    "updateCommentResponse": {
      "description": "Update Comment",
      "schema": {
        "$ref": "#/definitions/StatusResponse"
      }
    }
  },
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  }
}